backdoor: separate scripts depending on API availability
